+++ includes/structures.h
@@ -9,8 +9,8 @@ typedef struct	s_pt
 
 typedef struct	s_pos
 {
-	double	x;
-	double	y;
+	long double	x;
+	long double	y;
 }				t_pos;
 
 typedef struct	s_img
@@ -27,7 +27,7 @@ typedef struct	s_mlx
 	void	*mlx_ptr;
 	void	*win_ptr;
 	t_img	img;
-	double	scale;
+	long double	scale;
 	t_pos	pos;
 	int		max_ite;
 }				t_mlx;
